+++ extensions/mailsy/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Mailsy Changelog
 
+## [Changes] - 2024-10-25
+
+- Added `Logout` feature to log out of the account.
+- Removed the `Refresh` feature.
+
 ## [Features] - 2024-07-22
 
 - Functionality for opening an Account in the browser.